fpga-network/requirements.txt
# FPGA.Network Agent Dependencies

# Async HTTP
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0

# CLI
argparse

# Cryptography (for proofs)
# cryptography>=41.0.0

# Solana (for wallet integration)
# solana>=0.30.0
# solders>=0.18.0

# Optional: gRPC support
# grpcio>=1.60.0
# grpcio-tools>=1.60.0

# Development
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
